Add state machine

+++ sources/aliases.hpp
@@ -5,7 +5,7 @@
 namespace units{
 
     using Ms = uint32_t;
-    using Us = uint32_t;
+    using Us = uint64_t;
 
     inline constexpr Us Ms2Us(Ms millisecond){ return static_cast<Us>(millisecond) * 1000;}
     inline constexpr Ms Us2Ms(Us microsecond){ return static_cast<Ms>(microsecond / 1000);}
